requirements.txt
discord.py==1.2.5
requests==2.22.0